Got all preparer tests to pass locally and refactored and got another test to pass in handler
